--- jsonschema_testing/exceptions.py
@@ -1,20 +1,0 @@
-"""Exception classes used in jsonschema_testing.
-
-Copyright (c) 2020 Network To Code, LLC <info@networktocode.com>
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-
-
-class InvalidConfigAttribute(Exception):
-    """Exception raised if an invalid configuration parameter is specified"""
