+++ tests/fixtures/test_jsonschema/hostvars/can-vancouver-rt1/dns.yml
